requirements.txt
Django~=3.2
djangorestframework~=3.12.4
nltk~=3.5
numpy~=1.18.0
tensorflow~=2.3.0